Added ability to ignore a number of levels.
There are really two problems being solved here. Firstly UCM can't rename
across components, in a similiar way to Git with submodules. The same is
also true for renames across VOBs in either UCM or base Clearcase.

The second is that we might want to specify some of the git configuration
details in the root directory, perhaps such as .gitignore or .gitcc_config,
and we don't wish to add them to Clearcase.