Persist the since date to avoid duplicate commits.
This is really a quick-and-dirty fix. Ideally the storage or since is
done in such a way to be editable by users downstream of the daemon
without having access to it directly. It may be possible with annotated
tags and their timestamp, but may be annoying most of the time.