Avoid unnecessary rebase errors by comparing trees.
This is avoid rebase conflicts that result from a file modification
and then deletion in a single push from someone. As lshistory will
no longer return changes to files that are not visible master_cc
will diverge slightly from the original, and cause conflicts, even
though the final result is the same.

Where this will come unstuck is if there are actual changes upstream
in ClearCase as well. Ideally everyone is using Git anyway!