Fast import support
NOTE: This is not tested and is only a proof of concept.

I've been thinking about this since I started gitcc. The sad
thing is that I realised today that despite removing _some_ of
the dependency on having a non-bare repository, you _still_ need
it for rebases. I was hoping to share a singe repostiory
between multiple Clearcase repositories. The solution in any case
will be to have a file lock and only allow one gitcc daemon
to be rebasing at any one time.