Added rebaseTime configuration operation.
This lets you slow down the rate of rebasing. This is not really a good idea,
but I'm running 6+ daemons on my machine and I don't want it to chew up all
my resources. That combined with the fact that no-one is actually delivering
to our Clearcase streams _except_ Git means the rebase is really irrelevant.