Cleartool support
NOTE: Not tested yet!!!

This brings gitcc4j back into the territory of gitcc.
The reason for me doing this now is that CCRC is becoming
difficult to maintain with so many views. Hopefully by
using dynamic views the upkeep cost on my time will be much lower.
I'm hoping in the next few days to make cloneForUser()
create the views on the fly, joining streams if necessarily.